Add ability to change the class another class is indexed as
When indexing a class, the document builder will contain a term that
points to the type of object being indexed (this is named
`_hibernate_class`). We desired the need to change this as we were using
a DTO class to improve the speed it takes to index a collection of
objects retrieved from the database.

`Indexed` attribute now has a `HibernateClass` property that can be used
to define what class to use to determine the value of the
`_hibernate_class` term.